feat(#28): move set_section_wazuh_conf to configuration module
